Organize Rational code into Q
